requirements.txt
fastapi[standard]
sqlalchemy
pydantic
pyjwt[crypto]
passlib[bcrypt]
